cloud_worker/requirements.txt
conductor-python
moviepy